+++ app/src/main/java/org/simple/clinic/instantsearch/InstantSearchUi.kt
@@ -1,0 +1,6 @@
+package org.simple.clinic.instantsearch
+
+interface InstantSearchUi {
+  fun showSearchProgress()
+  fun hideSearchProgress()
+}